Move code to follow platformized projects, fix static analysis failures
