+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/excludes/factories/Intersections.java
@@ -16,6 +16,7 @@
 package org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories;
 
 import com.google.common.collect.Sets;
+import org.gradle.api.NonNullApi;
 import org.gradle.api.artifacts.ModuleIdentifier;
 import org.gradle.api.internal.artifacts.DefaultModuleIdentifier;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.specs.ExcludeAnyOf;
@@ -28,106 +29,77 @@ import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.specs
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.specs.ModuleIdSetExclude;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.specs.ModuleSetExclude;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import static java.util.stream.Collectors.toSet;
 
+@NonNullApi
 class Intersections {
     private final ExcludeFactory factory;
+    private final List<Intersection<? extends ExcludeSpec, ? extends ExcludeSpec>> intersections = new ArrayList<>();
 
     public Intersections(ExcludeFactory factory) {
         this.factory = factory;
+
+        // For the Any intersections, be sure to add the more specific type first, so it gets used if applicable
+        intersections.add(new IntersectAnyWithAny(factory));
+        intersections.add(new IntersectAnyWithBaseSpec(factory));
+
+        intersections.add(new IntersectGroupWithGroup(factory));
+        intersections.add(new IntersectGroupWithModuleId(factory));
+        intersections.add(new IntersectGroupWithGroupSet(factory));
+        intersections.add(new IntersectGroupWithModuleIdSet(factory));
+        intersections.add(new IntersectGroupWithModule(factory));
+        intersections.add(new IntersectGroupWithModuleSet(factory));
+
+        intersections.add(new IntersectGroupSetWithGroupSet(factory));
+        intersections.add(new IntersectGroupSetWithModuleId(factory));
+        intersections.add(new IntersectGroupSetWithModuleIdSet(factory));
+
+        intersections.add(new IntersectModuleWithModule(factory));
+        intersections.add(new IntersectModuleWithModuleId(factory));
+        intersections.add(new IntersectModuleWithModuleSet(factory));
+        intersections.add(new IntersectModuleWithModuleIdSet(factory));
+        intersections.add(new IntersectModuleWithGroupSet(factory));
+
+        intersections.add(new IntersectModuleIdWithModuleId(factory));
+        intersections.add(new IntersectModuleIdWithModuleIdSet(factory));
+        intersections.add(new IntersectModuleIdWithModuleSet(factory));
+
+        intersections.add(new IntersectModuleIdSetWithModuleIdSet(factory));
+        intersections.add(new IntersectModuleIdSetWithModuleSet(factory));
+
+        intersections.add(new IntersectModuleSetWithModuleSet(factory));
+        intersections.add(new IntersectModuleSetWithGroupSet(factory));
     }
 
-    /**
-     * Tries to compute an intersection of 2 specs.
-     * The result MUST be a simplification, otherwise this method returns null.
-     */
     ExcludeSpec tryIntersect(ExcludeSpec left, ExcludeSpec right) {
         if (left.equals(right)) {
             return left;
+        } else {
+            return intersections.stream()
+                .filter(i -> i.applies(left, right))
+                .findFirst()
+                .map(i -> i.intersect(left, right, factory))
+                .orElse(null);
         }
-
-        // Handle anyOf first because we don't want to special case it in
-        // every other case
-
-        if (left instanceof ExcludeAnyOf) {
-            ExcludeSpec excludeSpec = intersectAnyOf((ExcludeAnyOf) left, right);
-            if (excludeSpec != null) {
-                return excludeSpec;
-            }
-        } else if (right instanceof ExcludeAnyOf) {
-            ExcludeSpec excludeSpec = intersectAnyOf((ExcludeAnyOf) right, left);
-            if (excludeSpec != null) {
-                return excludeSpec;
-            }
-        }
-
-        /*
-         * The following cases are roughly ordered by the frequency of occurrence.  The order
-         * these checks are performed if CRITICAL, as not all of the intersectXYZ methods consider
-         * every type possibility for the right hand side operand.  Instead, some of these methods
-         * assume they will be called in the order here, and thus that certain types will not ever
-         * be supplied as their RHS operands.
-         *
-         * If the order of these checks is ever changed, then each intersectXYZ method must be
-         * updated to handle a RHS operand in all possible subtypes.  There are currently 6 types
-         * that extends ExcludeSpec and are relevant (GroupExclude, GroupSetExclude, ModuleExclude, ModuleIdExclude,
-         * ModuleIdSetExclude, ModuleSetExclude).  ExcludeEverything and ExcludeNothing are special
-         * cases handled above, and ArtifactExclude and CompositeExclude are also not relevant here.
-         */
-        if (left instanceof GroupExclude) {
-            return intersectGroup((GroupExclude) left, right);
-        } else if (right instanceof GroupExclude) {
-            return intersectGroup((GroupExclude) right, left);
-        } else if (left instanceof ModuleExclude) {
-            return intersectModule((ModuleExclude) left, right);
-        } else if (right instanceof ModuleExclude) {
-            return intersectModule((ModuleExclude) right, left);
-        } else if (left instanceof GroupSetExclude) {
-            return intersectGroupSet((GroupSetExclude) left, right);
-        } else if (right instanceof GroupSetExclude) {
-            return intersectGroupSet((GroupSetExclude) right, left);
-        } else if (left instanceof ModuleIdExclude) {
-            return intersectModuleId((ModuleIdExclude) left, right);
-        } else if (right instanceof ModuleIdExclude) {
-            return intersectModuleId((ModuleIdExclude) right, left);
-        } else if (left instanceof ModuleIdSetExclude) {
-            return intersectModuleIdSet((ModuleIdSetExclude) left, right);
-        } else if (right instanceof ModuleIdSetExclude) {
-            return intersectModuleIdSet((ModuleIdSetExclude) right, left);
-        } else if (left instanceof ModuleSetExclude) {
-            return intersectModuleSet((ModuleSetExclude) left, right);
-        } else if (right instanceof ModuleSetExclude) {
-            return intersectModuleSet((ModuleSetExclude) right, left);
-        }
-        return null;
     }
 
-    private ExcludeSpec intersectModuleSet(ModuleSetExclude left, ExcludeSpec right) {
-        if (right instanceof ModuleSetExclude) {
-            ModuleSetExclude msr = (ModuleSetExclude) right;
-            Set<String> modules = Sets.newHashSet(left.getModules());
-            modules.retainAll(msr.getModules());
-            if (modules.isEmpty()) {
-                return factory.nothing();
-            }
-            if (modules.size() == 1) {
-                return factory.module(modules.iterator().next());
-            }
-            return factory.moduleSet(modules);
+    private final class IntersectAnyWithAny extends AbstractIntersection<ExcludeAnyOf, ExcludeAnyOf> {
+        public IntersectAnyWithAny(ExcludeFactory factory) {
+            super(ExcludeAnyOf.class, ExcludeAnyOf.class, factory);
         }
-        return null;
-    }
 
-    private ExcludeSpec intersectAnyOf(ExcludeAnyOf left, ExcludeSpec right) {
-        Set<ExcludeSpec> leftComponents = left.getComponents();
-        if (right instanceof ExcludeAnyOf) {
-            Set<ExcludeSpec> rightComponents = ((ExcludeAnyOf) right).getComponents();
+        @Override
+        public ExcludeSpec doIntersect(ExcludeAnyOf left, ExcludeAnyOf right, ExcludeFactory factory) {
+            Set<ExcludeSpec> leftComponents = left.getComponents();
+            Set<ExcludeSpec> rightComponents = right.getComponents();
             Set<ExcludeSpec> common = Sets.newHashSet(leftComponents);
             common.retainAll(rightComponents);
-            if (common.size() >= 1) {
-                ExcludeSpec alpha = asUnion(common);
+            if (!common.isEmpty()) {
+                ExcludeSpec alpha = factory.fromUnion(common);
                 if (leftComponents.equals(common) || rightComponents.equals(common)) {
                     return alpha;
                 }
@@ -136,8 +108,8 @@ class Intersections {
                 Set<ExcludeSpec> remainderRight = Sets.newHashSet(rightComponents);
                 remainderRight.removeAll(common);
 
-                ExcludeSpec unionLeft = asUnion(remainderLeft);
-                ExcludeSpec unionRight = asUnion(remainderRight);
+                ExcludeSpec unionLeft = factory.fromUnion(remainderLeft);
+                ExcludeSpec unionRight = factory.fromUnion(remainderRight);
                 ExcludeSpec beta = factory.allOf(unionLeft, unionRight);
                 return factory.anyOf(alpha, beta);
             } else {
@@ -155,9 +127,19 @@ class Intersections {
                         }
                     }
                 }
-                return asUnion(intersections);
+                return factory.fromUnion(intersections);
             }
-        } else {
+        }
+    }
+
+    private final class IntersectAnyWithBaseSpec extends AbstractIntersection<ExcludeAnyOf, ExcludeSpec> {
+        private IntersectAnyWithBaseSpec(ExcludeFactory factory) {
+            super(ExcludeAnyOf.class, ExcludeSpec.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ExcludeAnyOf left, ExcludeSpec right, ExcludeFactory factory) {
+            Set<ExcludeSpec> leftComponents = left.getComponents();
             // Here, we will distribute A ∩ (B ∪ C) if, and only if, at
             // least one of the distribution operations (A ∩ B) can be simplified
             ExcludeSpec[] excludeSpecs = leftComponents.toArray(new ExcludeSpec[0]);
@@ -184,53 +166,277 @@ class Intersections {
                         simplified.add(factory.allOf(excludeSpecs[i], right));
                     }
                 }
-                return asUnion(simplified);
+                return factory.fromUnion(simplified);
+            } else {
+                return null;
             }
         }
-        return null;
+
+        @Override
+        public boolean applies(ExcludeSpec left, ExcludeSpec right) {
+            // We want to use the more specific AnyWithAny intersection if possible
+            return (left instanceof ExcludeAnyOf && !(right instanceof ExcludeAnyOf))
+                || (right instanceof ExcludeAnyOf && !(left instanceof ExcludeAnyOf));
+        }
     }
 
-    private ExcludeSpec asUnion(Set<ExcludeSpec> remainder) {
-        if (remainder.isEmpty()) {
-            // It's an intersection, and this method is always called on the remainder
-            // of a reduction operation. If the remainder is empty then it means that
-            // the intersection is empty
+    private final class IntersectGroupWithGroup extends AbstractIntersection<GroupExclude, GroupExclude> {
+        private IntersectGroupWithGroup(ExcludeFactory factory) {
+            super(GroupExclude.class, GroupExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupExclude left, GroupExclude right, ExcludeFactory factory) {
+            String group = left.getGroup();
+            // equality has been tested before so we know groups are different
             return factory.nothing();
         }
-        return remainder.size() == 1 ? remainder.iterator().next() : factory.anyOf(remainder);
     }
 
-    private ExcludeSpec intersectModuleId(ModuleIdExclude left, ExcludeSpec right) {
-        if (right instanceof ModuleIdExclude) {
-            if (left.equals(right)) {
+    private final class IntersectGroupWithModuleId extends AbstractIntersection<GroupExclude, ModuleIdExclude> {
+        private IntersectGroupWithModuleId(ExcludeFactory factory) {
+            super(GroupExclude.class, ModuleIdExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupExclude left, ModuleIdExclude right, ExcludeFactory factory) {
+            String group = left.getGroup();
+            if (right.getModuleId().getGroup().equals(group)) {
+                return right;
+            } else {
+                return factory.nothing();
+            }
+        }
+    }
+
+    private final class IntersectGroupWithGroupSet extends AbstractIntersection<GroupExclude, GroupSetExclude> {
+        private IntersectGroupWithGroupSet(ExcludeFactory factory) {
+            super(GroupExclude.class, GroupSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupExclude left, GroupSetExclude right, ExcludeFactory factory) {
+            String group = left.getGroup();
+            if (right.getGroups().stream().anyMatch(g -> g.equals(group))) {
                 return left;
             }
             return factory.nothing();
-        } else if (right instanceof ModuleIdSetExclude) {
-            Set<ModuleIdentifier> rightModuleIds = ((ModuleIdSetExclude) right).getModuleIds();
-            if (rightModuleIds.contains(left.getModuleId())) {
-                return left;
+        }
+    }
+
+    private final class IntersectGroupWithModuleIdSet extends AbstractIntersection<GroupExclude, ModuleIdSetExclude> {
+        private IntersectGroupWithModuleIdSet(ExcludeFactory factory) {
+            super(GroupExclude.class, ModuleIdSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupExclude left, ModuleIdSetExclude right, ExcludeFactory factory) {
+            String group = left.getGroup();
+            Set<ModuleIdentifier> moduleIds = right.getModuleIds().stream().filter(id -> id.getGroup().equals(group)).collect(toSet());
+            return factory.fromModuleIds(moduleIds);
+        }
+    }
+
+    private final class IntersectGroupWithModule extends AbstractIntersection<GroupExclude, ModuleExclude> {
+        private IntersectGroupWithModule(ExcludeFactory factory) {
+            super(GroupExclude.class, ModuleExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupExclude left, ModuleExclude right, ExcludeFactory factory) {
+            String group = left.getGroup();
+            return factory.moduleId(DefaultModuleIdentifier.newId(left.getGroup(), right.getModule()));
+        }
+    }
+
+    private final class IntersectGroupWithModuleSet extends AbstractIntersection<GroupExclude, ModuleSetExclude> {
+        private IntersectGroupWithModuleSet(ExcludeFactory factory) {
+            super(GroupExclude.class, ModuleSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupExclude left, ModuleSetExclude right, ExcludeFactory factory) {
+            String group = left.getGroup();
+            return factory.moduleIdSet(right.getModules()
+                .stream()
+                .map(module -> DefaultModuleIdentifier.newId(left.getGroup(), module))
+                .collect(toSet())
+            );
+        }
+    }
+
+    private final class IntersectGroupSetWithGroupSet extends AbstractIntersection<GroupSetExclude, GroupSetExclude> {
+        private IntersectGroupSetWithGroupSet(ExcludeFactory factory) {
+            super(GroupSetExclude.class, GroupSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupSetExclude left, GroupSetExclude right, ExcludeFactory factory) {
+            Set<String> groups = left.getGroups();
+            Set<String> common = Sets.newHashSet(right.getGroups());
+            common.retainAll(groups);
+            return factory.fromGroups(common);
+        }
+    }
+
+    private final class IntersectGroupSetWithModuleId extends AbstractIntersection<GroupSetExclude, ModuleIdExclude> {
+        private IntersectGroupSetWithModuleId(ExcludeFactory factory) {
+            super(GroupSetExclude.class, ModuleIdExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupSetExclude left, ModuleIdExclude right, ExcludeFactory factory) {
+            Set<String> groups = left.getGroups();
+            if (groups.contains(right.getModuleId().getGroup())) {
+                return right;
             }
             return factory.nothing();
-        } else if (right instanceof ModuleSetExclude) {
-            ModuleSetExclude moduleSetExclude = (ModuleSetExclude) right;
-            if (moduleSetExclude.getModules().contains(left.getModuleId().getName())) {
+        }
+    }
+
+    private final class IntersectGroupSetWithModuleIdSet extends AbstractIntersection<GroupSetExclude, ModuleIdSetExclude> {
+        private IntersectGroupSetWithModuleIdSet(ExcludeFactory factory) {
+            super(GroupSetExclude.class, ModuleIdSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(GroupSetExclude left, ModuleIdSetExclude right, ExcludeFactory factory) {
+            Set<String> groups = left.getGroups();
+            Set<ModuleIdentifier> filtered = right.getModuleIds()
+                .stream()
+                .filter(id -> groups.contains(id.getGroup()))
+                .collect(toSet());
+            return factory.fromModuleIds(filtered);
+        }
+    }
+
+    private final class IntersectModuleWithModule extends AbstractIntersection<ModuleExclude, ModuleExclude> {
+        private IntersectModuleWithModule(ExcludeFactory factory) {
+            super(ModuleExclude.class, ModuleExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleExclude left, ModuleExclude right, ExcludeFactory factory) {
+            String module = left.getModule();
+            if (right.getModule().equals(module)) {
                 return left;
             } else {
                 return factory.nothing();
             }
         }
-        return null;
     }
 
-    private ExcludeSpec intersectModuleIdSet(ModuleIdSetExclude left, ExcludeSpec right) {
-        Set<ModuleIdentifier> moduleIds = left.getModuleIds();
-        if (right instanceof ModuleIdSetExclude) {
-            Set<ModuleIdentifier> common = Sets.newHashSet(((ModuleIdSetExclude) right).getModuleIds());
+    private final class IntersectModuleWithModuleId extends AbstractIntersection<ModuleExclude, ModuleIdExclude> {
+        private IntersectModuleWithModuleId(ExcludeFactory factory) {
+            super(ModuleExclude.class, ModuleIdExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleExclude left, ModuleIdExclude right, ExcludeFactory factory) {
+            String module = left.getModule();
+            if (right.getModuleId().getName().equals(module)) {
+                return right;
+            } else {
+                return factory.nothing();
+            }
+        }
+    }
+
+    private final class IntersectModuleWithModuleSet extends AbstractIntersection<ModuleExclude, ModuleSetExclude> {
+        private IntersectModuleWithModuleSet(ExcludeFactory factory) {
+            super(ModuleExclude.class, ModuleSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleExclude left, ModuleSetExclude right, ExcludeFactory factory) {
+            String module = left.getModule();
+            if (right.getModules().stream().anyMatch(g -> g.equals(module))) {
+                return left;
+            }
+            return factory.nothing();
+        }
+    }
+
+    private final class IntersectModuleWithModuleIdSet extends AbstractIntersection<ModuleExclude, ModuleIdSetExclude> {
+        private IntersectModuleWithModuleIdSet(ExcludeFactory factory) {
+            super(ModuleExclude.class, ModuleIdSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleExclude left, ModuleIdSetExclude right, ExcludeFactory factory) {
+            String module = left.getModule();
+            Set<ModuleIdentifier> common = right.getModuleIds().stream().filter(id -> id.getName().equals(module)).collect(toSet());
+            return factory.fromModuleIds(common);
+        }
+    }
+
+    private final class IntersectModuleIdWithModuleId extends AbstractIntersection<ModuleIdExclude, ModuleIdExclude> {
+        private IntersectModuleIdWithModuleId(ExcludeFactory factory) {
+            super(ModuleIdExclude.class, ModuleIdExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleIdExclude left, ModuleIdExclude right, ExcludeFactory factory) {
+            if (left.equals(right)) {
+                return left;
+            }
+            return factory.nothing();
+        }
+    }
+
+    private final class IntersectModuleIdWithModuleIdSet extends AbstractIntersection<ModuleIdExclude, ModuleIdSetExclude> {
+        private IntersectModuleIdWithModuleIdSet(ExcludeFactory factory) {
+            super(ModuleIdExclude.class, ModuleIdSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleIdExclude left, ModuleIdSetExclude right, ExcludeFactory factory) {
+            Set<ModuleIdentifier> rightModuleIds = right.getModuleIds();
+            if (rightModuleIds.contains(left.getModuleId())) {
+                return left;
+            }
+            return factory.nothing();
+        }
+    }
+
+    private final class IntersectModuleIdWithModuleSet extends AbstractIntersection<ModuleIdExclude, ModuleSetExclude> {
+        private IntersectModuleIdWithModuleSet(ExcludeFactory factory) {
+            super(ModuleIdExclude.class, ModuleSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleIdExclude left, ModuleSetExclude right, ExcludeFactory factory) {
+            if (right.getModules().contains(left.getModuleId().getName())) {
+                return left;
+            } else {
+                return factory.nothing();
+            }
+        }
+    }
+
+    private final class IntersectModuleIdSetWithModuleIdSet extends AbstractIntersection<ModuleIdSetExclude, ModuleIdSetExclude> {
+        private IntersectModuleIdSetWithModuleIdSet(ExcludeFactory factory) {
+            super(ModuleIdSetExclude.class, ModuleIdSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleIdSetExclude left, ModuleIdSetExclude right, ExcludeFactory factory) {
+            Set<ModuleIdentifier> moduleIds = left.getModuleIds();
+            Set<ModuleIdentifier> common = Sets.newHashSet(right.getModuleIds());
             common.retainAll(moduleIds);
-            return moduleIds(common);
-        } else if (right instanceof ModuleSetExclude) {
-            Set<String> modules = ((ModuleSetExclude) right).getModules();
+            return factory.fromModuleIds(common);
+        }
+    }
+
+    private final class IntersectModuleIdSetWithModuleSet extends AbstractIntersection<ModuleIdSetExclude, ModuleSetExclude> {
+        private IntersectModuleIdSetWithModuleSet(ExcludeFactory factory) {
+            super(ModuleIdSetExclude.class, ModuleSetExclude.class, factory);
+        }
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleIdSetExclude left, ModuleSetExclude right, ExcludeFactory factory) {
+            Set<ModuleIdentifier> moduleIds = left.getModuleIds();
+            Set<String> modules = right.getModules();
             Set<ModuleIdentifier> identifiers = moduleIds.stream()
                 .filter(e -> modules.contains(e.getName()))
                 .collect(toSet());
@@ -243,120 +449,46 @@ class Intersections {
                 return factory.moduleIdSet(identifiers);
             }
         }
-        return null;
     }
 
-    private ExcludeSpec moduleIds(Set<ModuleIdentifier> common) {
-        if (common.isEmpty()) {
-            return factory.nothing();
+    private final class IntersectModuleSetWithModuleSet extends AbstractIntersection<ModuleSetExclude, ModuleSetExclude> {
+        private IntersectModuleSetWithModuleSet(ExcludeFactory factory) {
+            super(ModuleSetExclude.class, ModuleSetExclude.class, factory);
         }
-        if (common.size() == 1) {
-            return factory.moduleId(common.iterator().next());
-        }
-        return factory.moduleIdSet(common);
-    }
 
-    private ExcludeSpec intersectGroup(GroupExclude left, ExcludeSpec right) {
-        String group = left.getGroup();
-        if (right instanceof GroupExclude) {
-            // equality has been tested before so we know groups are different
-            return factory.nothing();
-        } else if (right instanceof ModuleIdExclude) {
-            if (((ModuleIdExclude) right).getModuleId().getGroup().equals(group)) {
-                return right;
-            } else {
+        @Override
+        public ExcludeSpec doIntersect(ModuleSetExclude left, ModuleSetExclude right, ExcludeFactory factory) {
+            Set<String> modules = Sets.newHashSet(left.getModules());
+            modules.retainAll(right.getModules());
+            if (modules.isEmpty()) {
                 return factory.nothing();
             }
-        } else if (right instanceof GroupSetExclude) {
-            if (((GroupSetExclude) right).getGroups().stream().anyMatch(g -> g.equals(group))) {
-                return left;
+            if (modules.size() == 1) {
+                return factory.module(modules.iterator().next());
             }
-            return factory.nothing();
-        } else if (right instanceof ModuleIdSetExclude) {
-            Set<ModuleIdentifier> moduleIds = ((ModuleIdSetExclude) right).getModuleIds().stream().filter(id -> id.getGroup().equals(group)).collect(toSet());
-            return moduleIdSet(moduleIds);
-        } else if (right instanceof ModuleExclude) {
-            return factory.moduleId(DefaultModuleIdentifier.newId(left.getGroup(), ((ModuleExclude) right).getModule()));
-        } else if (right instanceof ModuleSetExclude) {
-            ModuleSetExclude moduleSet = (ModuleSetExclude) right;
-            return factory.moduleIdSet(moduleSet.getModules()
-                .stream()
-                .map(module -> DefaultModuleIdentifier.newId(left.getGroup(), module))
-                .collect(toSet())
-            );
+            return factory.moduleSet(modules);
         }
-        return null;
     }
 
-    private ExcludeSpec moduleIdSet(Set<ModuleIdentifier> moduleIds) {
-        if (moduleIds.isEmpty()) {
-            return factory.nothing();
+    private final class IntersectModuleSetWithGroupSet extends AbstractIntersection<ModuleSetExclude, GroupSetExclude> {
+        private IntersectModuleSetWithGroupSet(ExcludeFactory factory) {
+            super(ModuleSetExclude.class, GroupSetExclude.class, factory);
         }
-        if (moduleIds.size() == 1) {
-            return factory.moduleId(moduleIds.iterator().next());
+
+        @Override
+        public ExcludeSpec doIntersect(ModuleSetExclude left, GroupSetExclude right, ExcludeFactory factory) {
+            return factory.moduleIdSet(right.getGroups().stream().flatMap(group -> left.getModules().stream().map(module -> DefaultModuleIdentifier.newId(group, module))).collect(toSet()));
         }
-        return factory.moduleIdSet(moduleIds);
     }
 
-    private ExcludeSpec intersectGroupSet(GroupSetExclude left, ExcludeSpec right) {
-        Set<String> groups = left.getGroups();
-        if (right instanceof GroupSetExclude) {
-            Set<String> common = Sets.newHashSet(((GroupSetExclude) right).getGroups());
-            common.retainAll(groups);
-            return groupSet(common);
-        } else if (right instanceof ModuleIdExclude) {
-            if (groups.contains(((ModuleIdExclude) right).getModuleId().getGroup())) {
-                return right;
-            }
-            return factory.nothing();
-        } else if (right instanceof ModuleIdSetExclude) {
-            Set<ModuleIdentifier> filtered = ((ModuleIdSetExclude) right).getModuleIds()
-                .stream()
-                .filter(id -> groups.contains(id.getGroup()))
-                .collect(toSet());
-            return moduleIdSet(filtered);
-        } else if (right instanceof  ModuleSetExclude) {
-            return factory.moduleIdSet(groups.stream().flatMap(group -> ((ModuleSetExclude) right).getModules().stream().map(module -> DefaultModuleIdentifier.newId(group, module))).collect(toSet()));
+    private final class IntersectModuleWithGroupSet extends AbstractIntersection<ModuleExclude, GroupSetExclude> {
+        private IntersectModuleWithGroupSet(ExcludeFactory factory) {
+            super(ModuleExclude.class, GroupSetExclude.class, factory);
         }
-        return null;
-    }
 
-    private ExcludeSpec groupSet(Set<String> common) {
-        if (common.isEmpty()) {
-            return factory.nothing();
+        @Override
+        public ExcludeSpec doIntersect(ModuleExclude left, GroupSetExclude right, ExcludeFactory factory) {
+            return factory.moduleIdSet(right.getGroups().stream().map(group -> DefaultModuleIdentifier.newId(group, left.getModule())).collect(toSet()));
         }
-        if (common.size() == 1) {
-            return factory.group(common.iterator().next());
-        }
-        return factory.groupSet(common);
-    }
-
-
-    private ExcludeSpec intersectModule(ModuleExclude left, ExcludeSpec right) {
-        String module = left.getModule();
-        if (right instanceof ModuleExclude) {
-            if (((ModuleExclude) right).getModule().equals(module)) {
-                return left;
-            } else {
-                return factory.nothing();
-            }
-        } else if (right instanceof ModuleIdExclude) {
-            if (((ModuleIdExclude) right).getModuleId().getName().equals(module)) {
-                return right;
-            } else {
-                return factory.nothing();
-            }
-        } else if (right instanceof ModuleSetExclude) {
-            if (((ModuleSetExclude) right).getModules().stream().anyMatch(g -> g.equals(module))) {
-                return left;
-            }
-            return factory.nothing();
-        } else if (right instanceof ModuleIdSetExclude) {
-            Set<ModuleIdentifier> common = ((ModuleIdSetExclude) right).getModuleIds().stream().filter(id -> id.getName().equals(module)).collect(toSet());
-            return moduleIdSet(common);
-        } else if (right instanceof GroupSetExclude) {
-            return factory.moduleIdSet(((GroupSetExclude) right).getGroups().stream().map(group -> DefaultModuleIdentifier.newId(group, module)).collect(toSet()));
-        }
-        return null;
     }
 }
