+++ subprojects/architecture-test/src/changes/archunit_store/internal-api-nullability.txt
@@ -138,7 +138,29 @@ Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.facto
 Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.CachingExcludeFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (CachingExcludeFactory.java:0)
 Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.DelegatingExcludeFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DelegatingExcludeFactory.java:0)
 Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.ExcludeFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ExcludeFactory.java:0)
-Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectAnyWithAny> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectAnyWithBaseSpec> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupSetWithGroupSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupSetWithModuleId> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupSetWithModuleIdSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupWithGroup> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupWithGroupSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupWithModule> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupWithModuleId> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupWithModuleIdSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectGroupWithModuleSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleIdSetWithModuleIdSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleIdSetWithModuleSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleIdWithModuleId> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleIdWithModuleIdSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleIdWithModuleSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleSetWithGroupSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleSetWithModuleSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleWithGroupSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleWithModule> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleWithModuleId> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleWithModuleIdSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
+Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.Intersections$IntersectModuleWithModuleSet> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Intersections.java:0)
 Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.LoggingExcludeFactory$Subject> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (LoggingExcludeFactory.java:0)
 Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.LoggingExcludeFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (LoggingExcludeFactory.java:0)
 Class <org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.factories.NormalizingExcludeFactory$FlattenOperationResult> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (NormalizingExcludeFactory.java:0)
