+++ subprojects/architecture-test/src/changes/archunit_store/public-api-mutable-properties.txt
@@ -346,13 +346,13 @@ Method <org.gradle.api.tasks.util.PatternSet.getAsExcludeSpec()> does not have r
 Method <org.gradle.api.tasks.util.PatternSet.getAsIncludeSpec()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.getAsSpec()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.getExcludeSpecs()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
-Method <org.gradle.api.tasks.util.PatternSet.getExcludes()> does not have raw return type assignable to org.gradle.api.provider.Property in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.getExcludeSpecsView()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
+Method <org.gradle.api.tasks.util.PatternSet.getExcludes()> does not have raw return type assignable to org.gradle.api.provider.Property in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.getExcludesView()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
-Method <org.gradle.api.tasks.util.PatternSet.getIncludeSpecsView()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
-Method <org.gradle.api.tasks.util.PatternSet.getIncludesView()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.getIncludeSpecs()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
+Method <org.gradle.api.tasks.util.PatternSet.getIncludeSpecsView()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.getIncludes()> does not have raw return type assignable to org.gradle.api.provider.Property in (PatternSet.java:0)
+Method <org.gradle.api.tasks.util.PatternSet.getIncludesView()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.isCaseSensitive()> does not have raw return type assignable to org.gradle.api.provider.Property in (PatternSet.java:0)
 Method <org.gradle.api.tasks.util.PatternSet.isEmpty()> does not have raw return type assignable to org.gradle.api.provider.Provider in (PatternSet.java:0)
 Method <org.gradle.caching.configuration.AbstractBuildCache.isEnabled()> does not have raw return type assignable to org.gradle.api.provider.Property in (AbstractBuildCache.java:0)
